[Loom] Isolate bytecode IR body serialization
Move IR allocation summaries, SSA value definitions, operation and
region encoding, and root-region payload tracking into a package-private
body component. The top-level writer now consumes body counts and region
ranges without owning their traversal.

The body writer retains streaming page output, exact arena
preallocation, and the existing definition-order SSA namespace. No
additional module walk or intermediate body representation is
introduced.
