+++ loom/src/loom/format/bytecode/BUILD.bazel
@@ -228,6 +228,7 @@ loom_cc_library(
         "//loom/src/loom/analysis:symbol_references",
         "//loom/src/loom/format:low_repr",
         "//loom/src/loom/format/bytecode/writer:attribute",
+        "//loom/src/loom/format/bytecode/writer:body",
         "//loom/src/loom/format/bytecode/writer:catalog",
         "//loom/src/loom/format/bytecode/writer:encoder",
         "//loom/src/loom/format/bytecode/writer:numbering",
