+++ loom/src/loom/format/bytecode/CMakeLists.txt
@@ -262,6 +262,7 @@ loom_cc_library(
     iree::io::stream
     loom::analysis::symbol_references
     loom::format::bytecode::writer::attribute
+    loom::format::bytecode::writer::body
     loom::format::bytecode::writer::catalog
     loom::format::bytecode::writer::encoder
     loom::format::bytecode::writer::numbering
